Group types and associated publishers
