+++ Sources/Player/Properties.swift
@@ -5,6 +5,7 @@
 //
 
 import AVFoundation
+import Combine
 
 /// Stream types.
 public enum StreamType {
@@ -64,5 +65,36 @@ public extension Player {
         static func empty(for player: AVPlayer) -> Self {
             Properties(playback: .empty, targetTime: nil)
         }
+    }
+
+    static func propertiesPublisher(for player: AVPlayer) -> AnyPublisher<Properties, Never> {
+        Publishers.CombineLatest(
+            playbackPublisher(for: player, queue: DispatchQueue(label: "ch.srgssr.pillarbox.player")),
+            seekTargetTimePublisher(for: player)
+        )
+        .map { Properties(playback: $0, targetTime: $1) }
+        .eraseToAnyPublisher()
+    }
+
+    static func playbackPublisher(for player: AVPlayer, queue: DispatchQueue) -> AnyPublisher<Properties.Playback, Never> {
+        periodicTimePublisher(for: player, interval: CMTimeMake(value: 1, timescale: 1), queue: queue)
+            .map { [weak player] time in
+                Properties.Playback(
+                    time: time,
+                    timeRange: Time.timeRange(for: player?.currentItem)
+                )
+            }
+            .eraseToAnyPublisher()
+    }
+
+    static func seekTargetTimePublisher(for player: AVPlayer) -> AnyPublisher<CMTime?, Never> {
+        Publishers.Merge(
+            NotificationCenter.default.weakPublisher(for: .willSeek, object: player)
+                .map { $0.userInfo?[SystemPlayer.SeekInfoKey.targetTime] as? CMTime },
+            NotificationCenter.default.weakPublisher(for: .didSeek, object: player)
+                .map { _ in nil }
+        )
+        .prepend(nil)
+        .eraseToAnyPublisher()
     }
 }
