+++ Tests/PlayerTests/PublisherTests.swift
@@ -112,6 +112,5 @@ final class PublisherTests: XCTestCase {
     // TODO:
     //  - Test without playing (no events; requires a way to check that values are never emitted)
     //  - Test with pause
-    //  - Similar individual tests for other publishers in Publishers.swift (Player package)
     //  - etc.
 }
